Refactor: Move authOptions and add CartProvider
