+++ package.json
@@ -2,10 +2,10 @@
   "name": "artisan-marketplace",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "type": "module",
     "start": "next start",
     "lint": "next lint"
   },
