+++ src/app/account/page.tsx
@@ -1,6 +1,6 @@
 // In src/app/account/page.tsx
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 import { redirect } from 'next/navigation';
 
 export default async function AccountPage() {
