+++ src/app/layout.tsx
@@ -3,17 +3,20 @@ import './globals.css';
 import { Header } from '@/components/Layout/Header';
 import Footer from '@/components/Layout/Footer';
 import AuthProvider from './providers/AuthProvider';
+import { CartProvider } from '@/context/CartContext';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
       <body>
-        <AuthProvider> {/* This is the new provider */}
-          <div className="min-h-screen flex flex-col">
-            <Header />
-            <main className="flex-grow">{children}</main>
-            <Footer />
-          </div>
+        <AuthProvider>
+          <CartProvider>
+            <div className="min-h-screen flex flex-col">
+              <Header />
+              <main className="flex-grow">{children}</main>
+              <Footer />
+            </div>
+          </CartProvider>
         </AuthProvider>
       </body>
     </html>
