+++ src/context/CartContext.tsx
@@ -2,7 +2,7 @@
 
 import React, { createContext, useContext, useState, ReactNode } from 'react'
 
-interface CartItem {
+export interface CartItem {
   id: number
   name: string
   price: number
