+++ tailwind.config.ts
@@ -1,6 +1,7 @@
-// In tailwind.config.js
-/** @type {import('tailwindcss').Config} */
-module.exports = {
+// In tailwind.config.ts
+import type { Config } from 'tailwindcss';
+
+const config: Config = {
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
@@ -26,3 +27,5 @@ module.exports = {
   },
   plugins: [],
 };
+
+export default config;
